LayoutTests/fast/mediastream/RTCPeerConnection-dtmf-expected.txt

Tests RTCDTMFSender.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.webkitGetUserMedia({audio:true}, gotStream, error); did not throw exception.
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 0
PASS pc is connected
PASS dtmfsender = pc.createDTMFSender(track); threw exception Error: SyntaxError: DOM Exception 12.
PASS dtmfsender = pc.createDTMFSender(track); did not throw exception.
PASS dtmfsender.canInsertDTMF is true
PASS dtmfsender.toneBuffer is "1"
PASS ontonechange was called.
PASS event.tone is "1"
PASS ontonechange was called.
PASS event.tone is ""
PASS successfullyParsed is true

TEST COMPLETE
